Share ZGC stub spilling logic across GCs
