+++ src/hotspot/cpu/aarch64/aarch64.ad
@@ -1126,10 +1126,6 @@ extern RegMask _NO_SPECIAL_REG32_mask;
 extern RegMask _NO_SPECIAL_REG_mask;
 extern RegMask _NO_SPECIAL_PTR_REG_mask;
 
-// Figure out which register class each belongs in: rc_int, rc_float or
-// rc_stack.
-enum RC { rc_bad, rc_int, rc_float, rc_predicate, rc_stack };
-
 class CallStubImpl {
 
   //--------------------------------------------------------------
