+++ src/hotspot/cpu/aarch64/gc/z/zBarrierSetAssembler_aarch64.hpp
@@ -187,12 +187,6 @@ public:
 #endif // COMPILER1
 
 #ifdef COMPILER2
-  OptoReg::Name encode_float_vector_register_size(const Node* node,
-                                                  OptoReg::Name opto_reg);
-
-  OptoReg::Name refine_register(const Node* node,
-                                OptoReg::Name opto_reg);
-
   void generate_c2_load_barrier_stub(MacroAssembler* masm,
                                      ZLoadBarrierStubC2* stub) const;
   void generate_c2_store_barrier_stub(MacroAssembler* masm,
