+++ src/hotspot/cpu/aarch64/register_aarch64.hpp
@@ -403,4 +403,7 @@ inline Register as_Register(FloatRegister reg) {
   return as_Register(reg->encoding());
 }
 
+// High-level register class of an OptoReg or a VMReg register.
+enum RC { rc_bad, rc_int, rc_float, rc_predicate, rc_stack };
+
 #endif // CPU_AARCH64_REGISTER_AARCH64_HPP
