+++ src/hotspot/cpu/ppc/gc/z/zBarrierSetAssembler_ppc.hpp
@@ -105,8 +105,6 @@ public:
 #endif // COMPILER1
 
 #ifdef COMPILER2
-  OptoReg::Name refine_register(const Node* node, OptoReg::Name opto_reg) const;
-
   void generate_c2_load_barrier_stub(MacroAssembler* masm, ZLoadBarrierStubC2* stub) const;
 
   void generate_c2_store_barrier_stub(MacroAssembler* masm, ZStoreBarrierStubC2* stub) const;
