+++ src/hotspot/cpu/riscv/gc/z/zBarrierSetAssembler_riscv.cpp
@@ -642,79 +642,8 @@ void ZBarrierSetAssembler::patch_barrier_relocation(address addr, int format) {
 
 #ifdef COMPILER2
 
-OptoReg::Name ZBarrierSetAssembler::refine_register(const Node* node, OptoReg::Name opto_reg) {
-  if (!OptoReg::is_reg(opto_reg)) {
-    return OptoReg::Bad;
-  }
-
-  const VMReg vm_reg = OptoReg::as_VMReg(opto_reg);
-  if (vm_reg->is_FloatRegister()) {
-    return opto_reg & ~1;
-  }
-
-  return opto_reg;
-}
-
 #undef __
 #define __ _masm->
-
-class ZSaveLiveRegisters {
-private:
-  MacroAssembler* const _masm;
-  RegSet                _gp_regs;
-  FloatRegSet           _fp_regs;
-  VectorRegSet          _vp_regs;
-
-public:
-  void initialize(ZBarrierStubC2* stub) {
-    // Record registers that needs to be saved/restored
-    RegMaskIterator rmi(stub->live());
-    while (rmi.has_next()) {
-      const OptoReg::Name opto_reg = rmi.next();
-      if (OptoReg::is_reg(opto_reg)) {
-        const VMReg vm_reg = OptoReg::as_VMReg(opto_reg);
-        if (vm_reg->is_Register()) {
-          _gp_regs += RegSet::of(vm_reg->as_Register());
-        } else if (vm_reg->is_FloatRegister()) {
-          _fp_regs += FloatRegSet::of(vm_reg->as_FloatRegister());
-        } else if (vm_reg->is_VectorRegister()) {
-          const VMReg vm_reg_base = OptoReg::as_VMReg(opto_reg & ~(VectorRegister::max_slots_per_register - 1));
-          _vp_regs += VectorRegSet::of(vm_reg_base->as_VectorRegister());
-        } else {
-          fatal("Unknown register type");
-        }
-      }
-    }
-
-    // Remove C-ABI SOE registers, tmp regs and _ref register that will be updated
-    if (stub->result() != noreg) {
-      _gp_regs -= RegSet::range(x18, x27) + RegSet::of(x2) + RegSet::of(x8, x9) + RegSet::of(x5, stub->result());
-    } else {
-      _gp_regs -= RegSet::range(x18, x27) + RegSet::of(x2, x5) + RegSet::of(x8, x9);
-    }
-  }
-
-  ZSaveLiveRegisters(MacroAssembler* masm, ZBarrierStubC2* stub)
-    : _masm(masm),
-      _gp_regs(),
-      _fp_regs(),
-      _vp_regs() {
-    // Figure out what registers to save/restore
-    initialize(stub);
-
-    // Save registers
-    __ push_reg(_gp_regs, sp);
-    __ push_fp(_fp_regs, sp);
-    __ push_v(_vp_regs, sp);
-  }
-
-  ~ZSaveLiveRegisters() {
-    // Restore registers
-    __ pop_v(_vp_regs, sp);
-    __ pop_fp(_fp_regs, sp);
-    __ pop_reg(_gp_regs, sp);
-  }
-};
 
 class ZSetupArguments {
 private:
@@ -781,7 +710,7 @@ void ZBarrierSetAssembler::generate_c2_load_barrier_stub(MacroAssembler* masm, Z
   }
 
   {
-    ZSaveLiveRegisters save_live_registers(masm, stub);
+    SaveLiveRegisters save_live_registers(masm, stub);
     ZSetupArguments setup_arguments(masm, stub);
     __ mv(t0, stub->slow_path());
     __ jalr(t0);
@@ -813,7 +742,7 @@ void ZBarrierSetAssembler::generate_c2_store_barrier_stub(MacroAssembler* masm, 
   __ bind(slow);
 
   {
-    ZSaveLiveRegisters save_live_registers(masm, stub);
+    SaveLiveRegisters save_live_registers(masm, stub);
     __ la(c_rarg0, stub->ref_addr());
 
     if (stub->is_native()) {
