+++ src/hotspot/cpu/x86/gc/shared/barrierSetAssembler_x86.hpp
@@ -28,7 +28,12 @@
 #include "asm/macroAssembler.hpp"
 #include "memory/allocation.hpp"
 #include "oops/access.hpp"
+#ifdef COMPILER2
+#include "opto/optoreg.hpp"
 
+class BarrierStubC2;
+class Node;
+#endif // COMPILER2
 class InterpreterMacroAssembler;
 
 class BarrierSetAssembler: public CHeapObj<mtGC> {
@@ -106,6 +111,63 @@ public:
   virtual void c2i_entry_barrier(MacroAssembler* masm);
 
   virtual void check_oop(MacroAssembler* masm, Register obj, Register tmp1, Register tmp2, Label& error);
+
+#ifdef COMPILER2
+  OptoReg::Name refine_register(const Node* node,
+                                OptoReg::Name opto_reg);
+#endif // COMPILER2
 };
+
+#ifdef COMPILER2
+
+#ifdef _LP64
+
+// This class saves and restores the registers that need to be preserved across
+// the runtime call represented by a given C2 barrier stub. Use as follows:
+// {
+//   SaveLiveRegisters save(masm, stub);
+//   ..
+//   __ call(RuntimeAddress(...);
+//   ..
+// }
+class SaveLiveRegisters {
+private:
+  struct XMMRegisterData {
+    XMMRegister _reg;
+    int         _size;
+
+    // Used by GrowableArray::find()
+    bool operator == (const XMMRegisterData& other) {
+      return _reg == other._reg;
+    }
+  };
+
+  MacroAssembler* const          _masm;
+  GrowableArray<Register>        _gp_registers;
+  GrowableArray<KRegister>       _opmask_registers;
+  GrowableArray<XMMRegisterData> _xmm_registers;
+  int                            _spill_size;
+  int                            _spill_offset;
+
+  static int xmm_compare_register_size(XMMRegisterData* left, XMMRegisterData* right);
+  static int xmm_slot_size(OptoReg::Name opto_reg);
+  static uint xmm_ideal_reg_for_size(int reg_size);
+  bool xmm_needs_vzeroupper() const;
+  void xmm_register_save(const XMMRegisterData& reg_data);
+  void xmm_register_restore(const XMMRegisterData& reg_data);
+  void gp_register_save(Register reg);
+  void opmask_register_save(KRegister reg);
+  void gp_register_restore(Register reg);
+  void opmask_register_restore(KRegister reg);
+  void initialize(BarrierStubC2* stub);
+
+public:
+  SaveLiveRegisters(MacroAssembler* masm, BarrierStubC2* stub);
+  ~SaveLiveRegisters();
+};
+
+#endif // _LP64
+
+#endif // COMPILER2
 
 #endif // CPU_X86_GC_SHARED_BARRIERSETASSEMBLER_X86_HPP
