+++ src/hotspot/cpu/x86/gc/z/zBarrierSetAssembler_x86.hpp
@@ -164,9 +164,6 @@ public:
 #endif // COMPILER1
 
 #ifdef COMPILER2
-  OptoReg::Name refine_register(const Node* node,
-                                OptoReg::Name opto_reg);
-
   void generate_c2_load_barrier_stub(MacroAssembler* masm,
                                      ZLoadBarrierStubC2* stub) const;
   void generate_c2_store_barrier_stub(MacroAssembler* masm,
