+++ src/hotspot/share/gc/z/c2/zBarrierSetC2.cpp
@@ -120,60 +120,22 @@ public:
 
 typedef ZArenaHashtable<intptr_t, bool, 4> ZOffsetTable;
 
-class ZBarrierSetC2State : public ArenaObj {
+class ZBarrierSetC2State : public BarrierSetC2State {
 private:
   GrowableArray<ZBarrierStubC2*>* _stubs;
-  Node_Array                      _live;
-  int                             _trampoline_stubs_count;
-  int                             _stubs_start_offset;
 
 public:
   ZBarrierSetC2State(Arena* arena)
-    : _stubs(new (arena) GrowableArray<ZBarrierStubC2*>(arena, 8,  0, nullptr)),
-      _live(arena),
-      _trampoline_stubs_count(0),
-      _stubs_start_offset(0) {}
+    : BarrierSetC2State(arena),
+      _stubs(new (arena) GrowableArray<ZBarrierStubC2*>(arena, 8,  0, nullptr)) {}
 
   GrowableArray<ZBarrierStubC2*>* stubs() {
     return _stubs;
   }
 
-  RegMask* live(const Node* node) {
-    if (!node->is_Mach()) {
-      // Don't need liveness for non-MachNodes
-      return nullptr;
-    }
-
-    const MachNode* const mach = node->as_Mach();
-    if (mach->barrier_data() == ZBarrierElided) {
-      // Don't need liveness data for nodes without barriers
-      return nullptr;
-    }
-
-    RegMask* live = (RegMask*)_live[node->_idx];
-    if (live == nullptr) {
-      live = new (Compile::current()->comp_arena()->AmallocWords(sizeof(RegMask))) RegMask();
-      _live.map(node->_idx, (Node*)live);
-    }
-
-    return live;
-  }
-
-  void inc_trampoline_stubs_count() {
-    assert(_trampoline_stubs_count != INT_MAX, "Overflow");
-    ++_trampoline_stubs_count;
-  }
-
-  int trampoline_stubs_count() {
-    return _trampoline_stubs_count;
-  }
-
-  void set_stubs_start_offset(int offset) {
-    _stubs_start_offset = offset;
-  }
-
-  int stubs_start_offset() {
-    return _stubs_start_offset;
+  bool needs_liveness_data(const MachNode* mach) const {
+    // Don't need liveness data for nodes without barriers
+    return mach->barrier_data() != ZBarrierElided;
   }
 };
 
@@ -201,30 +163,7 @@ int ZBarrierStubC2::stubs_start_offset() {
   return barrier_set_state()->stubs_start_offset();
 }
 
-ZBarrierStubC2::ZBarrierStubC2(const MachNode* node)
-  : _node(node),
-    _entry(),
-    _continuation() {}
-
-Register ZBarrierStubC2::result() const {
-  return noreg;
-}
-
-RegMask& ZBarrierStubC2::live() const {
-  return *barrier_set_state()->live(_node);
-}
-
-Label* ZBarrierStubC2::entry() {
-  // The _entry will never be bound when in_scratch_emit_size() is true.
-  // However, we still need to return a label that is not bound now, but
-  // will eventually be bound. Any eventually bound label will do, as it
-  // will only act as a placeholder, so we return the _continuation label.
-  return Compile::current()->output()->in_scratch_emit_size() ? &_continuation : &_entry;
-}
-
-Label* ZBarrierStubC2::continuation() {
-  return &_continuation;
-}
+ZBarrierStubC2::ZBarrierStubC2(const MachNode* node) : BarrierStubC2(node) {}
 
 ZLoadBarrierStubC2* ZLoadBarrierStubC2::create(const MachNode* node, Address ref_addr, Register ref) {
   AARCH64_ONLY(fatal("Should use ZLoadBarrierStubC2Aarch64::create"));
@@ -882,81 +821,6 @@ void ZBarrierSetC2::analyze_dominating_barriers() const {
   analyze_dominating_barriers_impl(loads, load_dominators);
   analyze_dominating_barriers_impl(stores, store_dominators);
   analyze_dominating_barriers_impl(atomics, atomic_dominators);
-}
-
-// == Reduced spilling optimization ==
-
-void ZBarrierSetC2::compute_liveness_at_stubs() const {
-  ResourceMark rm;
-  Compile* const C = Compile::current();
-  Arena* const A = Thread::current()->resource_area();
-  PhaseCFG* const cfg = C->cfg();
-  PhaseRegAlloc* const regalloc = C->regalloc();
-  RegMask* const live = NEW_ARENA_ARRAY(A, RegMask, cfg->number_of_blocks() * sizeof(RegMask));
-  ZBarrierSetAssembler* const bs = ZBarrierSet::assembler();
-  Block_List worklist;
-
-  for (uint i = 0; i < cfg->number_of_blocks(); ++i) {
-    new ((void*)(live + i)) RegMask();
-    worklist.push(cfg->get_block(i));
-  }
-
-  while (worklist.size() > 0) {
-    const Block* const block = worklist.pop();
-    RegMask& old_live = live[block->_pre_order];
-    RegMask new_live;
-
-    // Initialize to union of successors
-    for (uint i = 0; i < block->_num_succs; i++) {
-      const uint succ_id = block->_succs[i]->_pre_order;
-      new_live.OR(live[succ_id]);
-    }
-
-    // Walk block backwards, computing liveness
-    for (int i = block->number_of_nodes() - 1; i >= 0; --i) {
-      const Node* const node = block->get_node(i);
-
-      // Remove def bits
-      const OptoReg::Name first = bs->refine_register(node, regalloc->get_reg_first(node));
-      const OptoReg::Name second = bs->refine_register(node, regalloc->get_reg_second(node));
-      if (first != OptoReg::Bad) {
-        new_live.Remove(first);
-      }
-      if (second != OptoReg::Bad) {
-        new_live.Remove(second);
-      }
-
-      // Add use bits
-      for (uint j = 1; j < node->req(); ++j) {
-        const Node* const use = node->in(j);
-        const OptoReg::Name first = bs->refine_register(use, regalloc->get_reg_first(use));
-        const OptoReg::Name second = bs->refine_register(use, regalloc->get_reg_second(use));
-        if (first != OptoReg::Bad) {
-          new_live.Insert(first);
-        }
-        if (second != OptoReg::Bad) {
-          new_live.Insert(second);
-        }
-      }
-
-      // If this node tracks liveness, update it
-      RegMask* const regs = barrier_set_state()->live(node);
-      if (regs != nullptr) {
-        regs->OR(new_live);
-      }
-    }
-
-    // Now at block top, see if we have any changes
-    new_live.SUBTRACT(old_live);
-    if (new_live.is_NotEmpty()) {
-      // Liveness has refined, update and propagate to prior blocks
-      old_live.OR(new_live);
-      for (uint i = 1; i < block->num_preds(); ++i) {
-        Block* const pred = cfg->get_block_for_node(block->pred(i));
-        worklist.push(pred);
-      }
-    }
-  }
 }
 
 void ZBarrierSetC2::eliminate_gc_barrier(PhaseMacroExpand* macro, Node* node) const {
