+++ src/hotspot/share/gc/z/c2/zBarrierSetC2.hpp
@@ -41,12 +41,8 @@ class MachNode;
 
 class MacroAssembler;
 
-class ZBarrierStubC2 : public ArenaObj {
+class ZBarrierStubC2 : public BarrierStubC2 {
 protected:
-  const MachNode* _node;
-  Label           _entry;
-  Label           _continuation;
-
 static void register_stub(ZBarrierStubC2* stub);
 static void inc_trampoline_stubs_count();
 static int trampoline_stubs_count();
@@ -55,10 +51,6 @@ static int stubs_start_offset();
   ZBarrierStubC2(const MachNode* node);
 
 public:
-  RegMask& live() const;
-  Label* entry();
-  Label* continuation();
-
   virtual Register result() const = 0;
   virtual void emit_code(MacroAssembler& masm) = 0;
 };
@@ -108,7 +100,6 @@ public:
 
 class ZBarrierSetC2 : public BarrierSetC2 {
 private:
-  void compute_liveness_at_stubs() const;
   void analyze_dominating_barriers_impl(Node_List& accesses, Node_List& access_dominators) const;
   void analyze_dominating_barriers() const;
